requirements.txt
accelerate==0.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
bigdl-core-xe-21==2.7.0b20251015
bigdl-core-xe-addons-21==2.7.0b20251015
bigdl-core-xe-batch-21==2.7.0b20251015
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
datasets==4.4.1
dill==0.4.0
distro==1.9.0
dpcpp-cpp-rt==2024.0.2
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface==0.0.1
huggingface-hub==0.36.0
idna==3.11
intel-cmplr-lib-rt==2024.0.2
intel-cmplr-lic-rt==2024.0.2
intel-extension-for-pytorch==2.1.10+xpu
intel-opencl-rt==2024.0.2
intel-openmp==2024.0.2
ipex-llm==2.3.0b20251015
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
mkl==2024.0.0
mkl-dpcpp==2024.0.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.6.1
numpy==1.26.4
onemkl-sycl-blas==2024.0.0
onemkl-sycl-datafitting==2024.0.0
onemkl-sycl-dft==2024.0.0
onemkl-sycl-lapack==2024.0.0
onemkl-sycl-rng==2024.0.0
onemkl-sycl-sparse==2024.0.0
onemkl-sycl-stats==2024.0.0
onemkl-sycl-vm==2024.0.0
openai==2.11.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tabulate==0.9.0
tbb==2021.13.1
tokenizers==0.15.2
torch==2.1.0a0+cxx11.abi
torchvision==0.16.0a0+cxx11.abi
tqdm==4.67.1
transformers==4.37.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2 
urllib3==2.6.1
xxhash==3.6.0
yarl==1.22.0
-e .
